Move AppId type to public folder
