+++ packages/teams-js/src/public/index.ts
@@ -41,6 +41,7 @@ export {
   UrlDialogInfo,
 } from './interfaces';
 export { app } from './app';
+export { AppId } from './appId';
 export { appInstallDialog } from './appInstallDialog';
 export { barCode } from './barCode';
 export { chat, OpenGroupChatRequest, OpenSingleChatRequest } from './chat';
