+++ packages/teams-js/test/internal/appIdValidation.spec.ts
